backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.1
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
httpx==0.25.2
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.3
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.5
cryptography==46.0.5
cssselect==1.4.0
distlib==0.4.0
filelock==3.24.0
idna==3.11
lensfun==0.3.4
lockfile==0.12.2
lxml==6.0.2
msgpack==1.1.2
numpy==2.4.3
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pycairo==1.29.0
pycparser==2.23
PyGObject==3.54.5
pyserial==3.5
python-discovery==1.1.3
requests==2.32.5
scour==0.38.2
six==1.17.0
soupsieve==2.8.3
TBB==0.2
tinycss2==1.5.1
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
zstandard==0.25.0
Pyrogram==2.0.106
fastapi-cors==0.0.6
